fix(ci): add package-lock.json files to fix caching
The GitHub Actions workflow was failing because the 'setup-node' action could not find the 'package-lock.json' files for caching.

This was happening because 'package-lock.json' was in the '.gitignore' file. This change removes 'package-lock.json' from '.gitignore' and adds the lock files for the root, api, docs, and web-components packages to the repository.